# Speed up Git graph creation